Fixed bug where hex characters were being saved to char arrayinstead …

+++ P1/src/Utilizacion_API/os_API.h
@@ -23,10 +23,10 @@ typedef struct disk
 
 typedef struct osFILE {
     char name[28];
-    unsigned int index_ptr;
-    unsigned int directory_ptr;
-    unsigned int size;
-    unsigned int bytes_read;
+    unsigned long int index_ptr;
+    unsigned long int directory_ptr;
+    unsigned long int size;
+    unsigned long int bytes_read;
 }osFILE;
